feat: move agglayer-prover to provers repo
