+++ Cargo.toml
@@ -23,7 +23,6 @@ agglayer-config = { path = "crates/agglayer-config" }
 agglayer-contracts = { path = "crates/agglayer-contracts" }
 agglayer-node = { path = "crates/agglayer-node" }
 agglayer-primitives = { path = "crates/agglayer-primitives" }
-agglayer-prover = { path = "crates/agglayer-prover", features = ["testutils"] }
 agglayer-signer = { path = "crates/agglayer-signer" }
 agglayer-storage = { path = "crates/agglayer-storage" }
 agglayer-telemetry = { path = "crates/agglayer-telemetry" }
@@ -31,6 +30,13 @@ agglayer-types = { path = "crates/agglayer-types" }
 pessimistic-proof = { path = "crates/pessimistic-proof" }
 pessimistic-proof-core = { path = "crates/pessimistic-proof-core" }
 pessimistic-proof-test-suite = { path = "crates/pessimistic-proof-test-suite" }
+
+agglayer-prover = { git = "https://github.com/agglayer/provers.git", features = [
+    "testutils",
+] }
+
+agglayer-prover-config = { git = "https://github.com/agglayer/provers.git" }
+agglayer-prover-types = { git = "https://github.com/agglayer/provers.git" }
 
 # Core dependencies
 alloy = { version = "0.8.1", features = ["full"] }
@@ -82,3 +88,13 @@ test-log = "0.2.16"
 sp1-core-machine = "=4.0.0"
 sp1-sdk = "=4.0.0"
 sp1-prover = "=4.0.0"
+
+# [patch.'https://github.com/agglayer/provers.git']
+# agglayer-prover-config = { path = "../provers/crates/agglayer-prover-config" }
+# agglayer-prover-types = { path = "../provers/crates/agglayer-prover-types" }
+# agglayer-prover = { path = "../provers/crates/agglayer-prover" }
+#
+[patch.'https://github.com/agglayer/agglayer.git']
+pessimistic-proof-core = { path = "./crates/pessimistic-proof-core" }
+pessimistic-proof = { path = "./crates/pessimistic-proof" }
+agglayer-telemetry = { path = "./crates/agglayer-telemetry" }
