+++ crates/agglayer-aggregator-notifier/Cargo.toml
@@ -8,10 +8,12 @@ license.workspace = true
 agglayer-certificate-orchestrator = { path = "../agglayer-certificate-orchestrator" }
 agglayer-config = { path = "../agglayer-config" }
 agglayer-contracts = { path = "../agglayer-contracts" }
-agglayer-prover-types = { path = "../agglayer-prover-types" }
 agglayer-storage = { path = "../agglayer-storage" }
 agglayer-types.workspace = true
 pessimistic-proof = { path = "../pessimistic-proof" }
+
+agglayer-prover-types.workspace = true
+agglayer-prover-config.workspace = true
 
 anyhow.workspace = true
 async-trait.workspace = true
@@ -32,7 +34,7 @@ sp1-sdk.workspace = true
 sp1-prover.workspace = true
 
 [dev-dependencies]
-agglayer-prover = { path = "../agglayer-prover", features = ["testutils"] }
+agglayer-prover = { workspace = true, features = ["testutils"] }
 agglayer-storage = { path = "../agglayer-storage", features = ["testutils"] }
 pessimistic-proof-test-suite = { path = "../pessimistic-proof-test-suite" }
