+++ crates/agglayer-config/tests/validate_deserialize.rs
@@ -1,7 +1,7 @@
 use std::path::Path;
 
-use agglayer_config::prover::{AgglayerProverType, CpuProverConfig, NetworkProverConfig};
-use agglayer_config::{prover::ProverConfig, Config};
+use agglayer_config::Config;
+use agglayer_prover_config::ProverConfig;
 use insta::assert_toml_snapshot;
 use pretty_assertions::assert_eq;
 
@@ -88,60 +88,4 @@ fn prover_grpc_max_decoding_message_size() {
     });
 
     assert_eq!(config.grpc.max_decoding_message_size, 100 * 1024 * 1024);
-}
-
-#[test]
-fn network_prover() {
-    let input = "./tests/fixtures/valide_config/prover_config_network_prover.toml";
-    let config = ProverConfig::try_load(Path::new(input)).unwrap();
-
-    assert_eq!(
-        config.primary_prover,
-        AgglayerProverType::NetworkProver(NetworkProverConfig {
-            proving_request_timeout: Some(std::time::Duration::from_secs(300)),
-            proving_timeout: std::time::Duration::from_secs(600),
-        })
-    );
-
-    assert_eq!(config.fallback_prover, None);
-}
-
-#[test]
-fn cpu_prover() {
-    let input = "./tests/fixtures/valide_config/prover_config_cpu_prover.toml";
-    let config = ProverConfig::try_load(Path::new(input)).unwrap();
-
-    assert_eq!(
-        config.primary_prover,
-        AgglayerProverType::CpuProver(CpuProverConfig {
-            max_concurrency_limit: 10,
-            proving_request_timeout: Some(std::time::Duration::from_secs(300)),
-            proving_timeout: std::time::Duration::from_secs(600),
-        })
-    );
-
-    assert_eq!(config.fallback_prover, None);
-}
-
-#[test]
-fn network_and_cpu_prover() {
-    let input = "./tests/fixtures/valide_config/prover_config_primary_fallback_prover.toml";
-    let config = ProverConfig::try_load(Path::new(input)).unwrap();
-
-    assert_eq!(
-        config.primary_prover,
-        AgglayerProverType::NetworkProver(NetworkProverConfig {
-            proving_request_timeout: Some(std::time::Duration::from_secs(300)),
-            proving_timeout: std::time::Duration::from_secs(600),
-        })
-    );
-
-    assert_eq!(
-        config.fallback_prover,
-        Some(AgglayerProverType::CpuProver(CpuProverConfig {
-            max_concurrency_limit: 10,
-            proving_request_timeout: Some(std::time::Duration::from_secs(300)),
-            proving_timeout: std::time::Duration::from_secs(600),
-        }))
-    );
 }
