--- crates/agglayer-prover-types/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "agglayer-prover-types"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-anyhow.workspace = true
-bincode.workspace = true
-prost = "0.13.4"
-serde.workspace = true
-thiserror.workspace = true
-tonic = { workspace = true, default-features = false, features = [
-    "prost",
-    "codegen",
-    "transport",
-] }
-
-agglayer-types.workspace = true
-pessimistic-proof = { path = "../pessimistic-proof" }
-
-[build-dependencies]
-tonic-build = { version = "0.12", default-features = false, features = [
-    "prost",
-    "transport",
-] }
