--- crates/agglayer-prover-types/proto/v1/proof_generation.proto
@@ -1,20 +1,0 @@
-syntax = "proto3";
-
-package agglayer.prover.v1;
-
-service ProofGenerationService {
-    rpc GenerateProof(ProofGenerationRequest) returns (ProofGenerationResponse);
-}
-
-message ProofGenerationRequest {
-    bytes initial_state = 1;
-    bytes batch_header = 2;
-}
-
-message ProofGenerationResponse {
-    bytes proof = 1;
-}
-
-message ProofGenerationError {
-    bytes error = 1;
-}
