--- crates/agglayer-prover/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "agglayer-prover"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-anyhow.workspace = true
-bincode.workspace = true
-buildstructor.workspace = true
-futures.workspace = true
-thiserror.workspace = true
-tokio = { workspace = true, features = ["full"] }
-tokio-util.workspace = true
-toml.workspace = true
-tracing-subscriber = { workspace = true, features = ["env-filter", "json"] }
-tracing.workspace = true
-tower = { workspace = true, features = ["timeout"] }
-tonic = { workspace = true, features = ["zstd"] }
-tonic-health = "0.12.3"
-tonic-reflection = "0.12.3"
-
-agglayer-config = { path = "../agglayer-config" }
-agglayer-prover-types = { path = "../agglayer-prover-types" }
-agglayer-telemetry = { path = "../agglayer-telemetry" }
-pessimistic-proof = { path = "../pessimistic-proof" }
-agglayer-types.workspace = true
-
-sp1-sdk = { workspace = true, features = ["native-gnark"] }
-sp1-prover = { workspace = true, features = ["native-gnark"] }
-
-[features]
-default = []
-testutils = []
-gpu = ["sp1-sdk/cuda"]
