+++ crates/agglayer/Cargo.toml
@@ -11,7 +11,8 @@ dotenvy.workspace = true
 toml.workspace = true
 
 agglayer-node = { path = "../agglayer-node" }
-agglayer-prover = { path = "../agglayer-prover" }
+agglayer-prover = { workspace = true }
+agglayer-prover-config = { workspace = true }
 agglayer-config = { path = "../agglayer-config" }
 
 [dev-dependencies]
