+++ crates/pessimistic-proof/Cargo.toml
@@ -13,6 +13,7 @@ pessimistic-proof-core.workspace = true
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1", features = ["arbitrary_precision"] }
 serde_with = { version = "3" }
+sp1-sdk = { workspace = true, optional = true }
 thiserror.workspace = true
 tiny-keccak = { version = "2.0", features = ["keccak"] }
 
@@ -23,3 +24,7 @@ rs_merkle = { version = "1.4", default-features = false }
 
 [lints]
 workspace = true
+
+[features]
+default = ["sp1"]
+sp1 = ["dep:sp1-sdk"]
