+++ crates/pessimistic-proof/src/error.rs
@@ -1,0 +1,39 @@
+use serde::{Deserialize, Serialize};
+use sp1_sdk::SP1VerificationError;
+
+#[derive(Clone, Debug, Serialize, Deserialize, thiserror::Error, PartialEq, Eq)]
+pub enum ProofVerificationError {
+    #[error("Version mismatch: {0}")]
+    VersionMismatch(String),
+    #[error("Core machine verification error: {0}")]
+    Core(String),
+    #[error("Recursion verification error: {0}")]
+    Recursion(String),
+    #[error("Plonk verification error: {0}")]
+    Plonk(String),
+    #[error("Groth16 verification error: {0}")]
+    Groth16(String),
+    #[error("Invalid public values")]
+    InvalidPublicValues,
+}
+
+impl From<SP1VerificationError> for ProofVerificationError {
+    fn from(err: SP1VerificationError) -> Self {
+        match err {
+            SP1VerificationError::VersionMismatch(version) => {
+                ProofVerificationError::VersionMismatch(version)
+            }
+            SP1VerificationError::Core(core) => ProofVerificationError::Core(core.to_string()),
+            SP1VerificationError::Recursion(recursion) => {
+                ProofVerificationError::Recursion(recursion.to_string())
+            }
+            SP1VerificationError::Plonk(error) => ProofVerificationError::Plonk(error.to_string()),
+            SP1VerificationError::Groth16(error) => {
+                ProofVerificationError::Groth16(error.to_string())
+            }
+            SP1VerificationError::InvalidPublicValues => {
+                ProofVerificationError::InvalidPublicValues
+            }
+        }
+    }
+}
