+++ crates/pessimistic-proof/src/proof.rs
@@ -1,4 +1,13 @@
 pub use pessimistic_proof_core::PessimisticProofOutput;
+use pessimistic_proof_core::{
+    local_exit_tree::hasher::Keccak256Hasher, multi_batch_header::MultiBatchHeader, NetworkState,
+};
+use serde::{Deserialize, Serialize};
+use sp1_sdk::{
+    Prover, ProverClient, SP1Proof, SP1ProofWithPublicValues, SP1PublicValues, SP1Stdin,
+};
+
+use crate::ELF;
 
 pub trait DisplayToHex {
     fn display_to_hex(&self) -> String;
@@ -18,6 +27,39 @@ impl DisplayToHex for PessimisticProofOutput {
             self.new_local_exit_root,
             self.new_pessimistic_root,
         )
+    }
+}
+
+/// Proof is a wrapper around all the different types of proofs that can be
+/// generated
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum Proof {
+    SP1(SP1ProofWithPublicValues),
+}
+
+impl Proof {
+    pub fn dummy() -> Self {
+        Self::SP1(SP1ProofWithPublicValues {
+            proof: SP1Proof::Core(vec![]),
+            public_values: SP1PublicValues::new(),
+            sp1_version: "".to_string(),
+        })
+    }
+
+    pub fn new_for_test(
+        state: &NetworkState,
+        multi_batch_header: &MultiBatchHeader<Keccak256Hasher>,
+    ) -> Self {
+        let mock = ProverClient::builder().mock().build();
+        let (p, _v) = mock.setup(ELF);
+
+        let mut stdin = SP1Stdin::new();
+        stdin.write(state);
+        stdin.write(multi_batch_header);
+
+        let proof = mock.prove(&p, &stdin).plonk().run().unwrap();
+
+        Proof::SP1(proof)
     }
 }
